+++ lib/src/visitor/evaluate.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_evaluate.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: 72aeb044260f1f64f121280996c97a762b7313b0
+// Checksum: 0f55a90c029d7b2041baf51afba2694bc1aa115d
 //
 // ignore_for_file: unused_import
 
@@ -58,6 +58,7 @@ import '../value.dart';
 import 'expression_to_calc.dart';
 import 'interface/css.dart';
 import 'interface/expression.dart';
+import 'interface/if_condition_expression.dart';
 import 'interface/modifiable_css.dart';
 import 'interface/statement.dart';
 import 'serialize.dart';
@@ -143,6 +144,7 @@ final class _EvaluateVisitor
     implements
         StatementVisitor<Value?>,
         ExpressionVisitor<Value>,
+        IfConditionExpressionVisitor<Object /* String | bool */ >,
         CssVisitor<void> {
   /// The import cache used to import other stylesheets.
   final ImportCache? _importCache;
@@ -2569,7 +2571,7 @@ final class _EvaluateVisitor
   /// necessary if [condition] is also a [SupportsOperation].
   String _parenthesize(
     SupportsCondition condition, [
-    String? operator,
+    BooleanOperator? operator,
   ]) {
     switch (condition) {
       case SupportsNegation():
@@ -2833,6 +2835,88 @@ final class _EvaluateVisitor
 
   SassBoolean visitBooleanExpression(BooleanExpression node) =>
       SassBoolean(node.value);
+
+  Value visitIfExpression(IfExpression node) {
+    List<(String, Value)>? results;
+    for (var (condition, expression) in node.branches) {
+      var result = condition?.accept(this) ?? true;
+      var value = expression.accept(this);
+      switch (result) {
+        case String condition:
+          results ??= [];
+          results.add((condition, value));
+
+        case true when results != null:
+          results.add(('else', value));
+
+        case true:
+          return value;
+      }
+    }
+
+    if (results == null) return sassNull;
+    return SassString(
+        'if(' +
+            results.map((pair) => '${pair.$1}: ${pair.$2}').join('; ') +
+            ')',
+        quotes: false);
+  }
+
+  Object /* String | bool */ visitIfConditionParenthesized(
+          IfConditionParenthesized node) =>
+      switch (node.expression.accept(this)) {
+        String string => '($string)',
+        var result => result,
+      };
+
+  Object /* String | bool */ visitIfConditionNegation(
+          IfConditionNegation node) =>
+      switch (node.expression.accept(this)) {
+        String string => 'not $string',
+        bool result => !result,
+        _ => throw UnsupportedError('unreachable'),
+      };
+
+  Object /* String | bool */ visitIfConditionOperation(
+      IfConditionOperation node) {
+    List<(IfConditionExpression, String)>? values;
+    for (var expression in node.expressions) {
+      switch (expression.accept(this)) {
+        case String right:
+          values ??= [];
+          values.add((expression, right));
+        case false when node.op == BooleanOperator.and:
+          return false;
+        case true when node.op == BooleanOperator.or:
+          return true;
+      }
+    }
+
+    return switch (values) {
+      null => node.op == BooleanOperator.and,
+
+      // If the only CSS node left in the operation is parenthesized, remove
+      // the parentheses. This is guaranteed to be valid because parentheses
+      // contain an `<if-group>` and this operation is itself an
+      // `<if-group>`.
+      [(IfConditionParenthesized(), var string)] =>
+        string.substring(1, string.length - 1),
+      _ => values.map((pair) => pair.$2).join(' ${node.op} '),
+    };
+  }
+
+  Object /* String | bool */ visitIfConditionFunction(
+          IfConditionFunction node) =>
+      _performInterpolation(node.name) +
+      '(' +
+      _performInterpolation(node.arguments) +
+      ')';
+
+  Object /* String | bool */ visitIfConditionSass(IfConditionSass node) =>
+      node.expression.accept(this).isTruthy;
+
+  Object /* String | bool */ visitIfConditionRaw(IfConditionRaw node) =>
+      _performInterpolation(node.text);
 
   Value visitLegacyIfExpression(LegacyIfExpression node) {
     var (positional, named) = _evaluateMacroArguments(node);